feat: add createCodec() factory for type-safe msgpack
- createCodec(schema) returns { encode, decode, encodeSafe, decodeSafe }
  - decode/encode: typed but not validated (fast path for SDK)
  - decodeSafe/encodeSafe: validates with Zod (safe path for Spire trust boundary)
- Extract shared Packr config + browser-safe encode wrapper into src/codec.ts
- Client.ts now imports msgpack from codec.ts instead of inline Packr
- Export createCodec + msgpack from package public API
- 28/28 E2E tests pass
